Renamed san_init as san_elements
