+++ assets/js/openings/san_movetext/index.js
@@ -1,4 +1,4 @@
-import { ws, openingsSanMovetext } from '../../../js/base/san_init.js';
+import { ws, openingsSanMovetext } from '../../../js/base/san_elements.js';
 
 await ws.connect();
