+++ assets/js/play/computer/index.js
@@ -1,4 +1,4 @@
-import { ws, playComputer } from '../../../js/base/san_init.js';
+import { ws, playComputer } from '../../../js/base/san_elements.js';
 
 await ws.connect();
