+++ assets/js/play/friend/enter_invite_code.js
@@ -1,4 +1,4 @@
-import { ws, enterInviteCode } from '../../../js/base/san_init.js';
+import { ws, enterInviteCode } from '../../../js/base/san_elements.js';
 
 await ws.connect();
