+++ assets/js/play/friend/index.js
@@ -1,4 +1,4 @@
-import { ws, copyInviteCode, playFriend } from '../../../js/base/san_init.js';
+import { ws, copyInviteCode, playFriend } from '../../../js/base/san_elements.js';
 
 await ws.connect();
